Split web endpoint API tests
